Pass validation errors from unordered content generation up to group
As both `ModelVisitor.iter_unordered_content()` and
`XsdGroup.iter_encode()` can move the model position along, we have to
be able to record validation errors at both levels or we allow
validation errors to be consumed without appropriate handling.

A different solution would be to iterate through a `ModelVisitor` to
generate the content and iterate a separate `ModelVisitor` for the
validation.